refactor(temporal): extract delete tracks tests and rename add tracks tests file
