+++ packages/temporal/mpe/workflows/mpe_add_tracks_test.go
@@ -14,11 +14,11 @@ import (
 	"go.temporal.io/sdk/workflow"
 )
 
-type EditingPlaylistTestSuite struct {
+type AddTracksTestSuite struct {
 	UnitTestSuite
 }
 
-func (s *EditingPlaylistTestSuite) Test_AddTracks() {
+func (s *AddTracksTestSuite) Test_AddTracks() {
 	initialTracksIDs := []string{
 		faker.UUIDHyphenated(),
 	}
@@ -123,7 +123,7 @@ func (s *EditingPlaylistTestSuite) Test_AddTracks() {
 	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
 }
 
-func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistBeforeFetchingInformationFails() {
+func (s *AddTracksTestSuite) Test_AddingTrackAlreadyInPlaylistBeforeFetchingInformationFails() {
 	initialTracksIDs := []string{
 		faker.UUIDHyphenated(),
 	}
@@ -203,7 +203,7 @@ func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistBeforeFetchi
 	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
 }
 
-func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchingInformationSucceedsIfNotAllTracksAreDuplicated() {
+func (s *AddTracksTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchingInformationSucceedsIfNotAllTracksAreDuplicated() {
 	initialTracksIDs := []string{
 		faker.UUIDHyphenated(),
 	}
@@ -346,7 +346,7 @@ func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchin
 	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
 }
 
-func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchingInformationFailsIfAllTracksAreDuplicated() {
+func (s *AddTracksTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchingInformationFailsIfAllTracksAreDuplicated() {
 	initialTracksIDs := []string{
 		faker.UUIDHyphenated(),
 	}
@@ -493,245 +493,6 @@ func (s *EditingPlaylistTestSuite) Test_AddingTrackAlreadyInPlaylistAfterFetchin
 	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
 }
 
-func (s *EditingPlaylistTestSuite) Test_DeleteTracksAndSendAcknowledgement() {
-	initialTracksIDs := []string{
-		faker.UUIDHyphenated(),
-		faker.UUIDHyphenated(),
-	}
-	params, roomCreatorDeviceID := s.getWorkflowInitParams(initialTracksIDs)
-	var a *activities_mpe.Activities
-
-	initialTracksMetadata := []shared.TrackMetadata{
-		{
-			ID:         initialTracksIDs[0],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-		{
-			ID:         initialTracksIDs[1],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-	}
-
-	tick := 1 * time.Millisecond
-	resetMock, registerDelayedCallbackWrapper := s.initTestEnv()
-
-	defer resetMock()
-
-	// Common activities calls
-	s.env.OnActivity(
-		a.MpeCreationAcknowledgementActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Once()
-	s.env.OnActivity(
-		activities.FetchTracksInformationActivity,
-		mock.Anything,
-		initialTracksIDs,
-	).Return(initialTracksMetadata, nil).Once()
-
-	// Specific activities calls
-	s.env.OnActivity(
-		a.AcknowledgeDeletingTracksActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Once()
-
-	initialTracksFetched := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Equal(initialTracksMetadata, mpeState.Tracks)
-	}, initialTracksFetched)
-
-	deleteTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		s.emitDeleteTracksSignal(shared_mpe.NewDeleteTracksSignalArgs{
-			TracksIDs: initialTracksIDs,
-			UserID:    params.RoomCreatorUserID,
-			DeviceID:  roomCreatorDeviceID,
-		})
-	}, deleteTracks)
-
-	checkDeletedTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Empty(mpeState.Tracks)
-	}, checkDeletedTracks)
-
-	s.env.ExecuteWorkflow(MpeRoomWorkflow, params)
-
-	s.True(s.env.IsWorkflowCompleted())
-	err := s.env.GetWorkflowError()
-	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
-}
-
-func (s *EditingPlaylistTestSuite) Test_PreventUnknownUserFromDeletingTracks() {
-	initialTracksIDs := []string{
-		faker.UUIDHyphenated(),
-		faker.UUIDHyphenated(),
-	}
-	params, _ := s.getWorkflowInitParams(initialTracksIDs)
-	var a *activities_mpe.Activities
-
-	initialTracksMetadata := []shared.TrackMetadata{
-		{
-			ID:         initialTracksIDs[0],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-		{
-			ID:         initialTracksIDs[1],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-	}
-
-	tick := 1 * time.Millisecond
-	resetMock, registerDelayedCallbackWrapper := s.initTestEnv()
-
-	defer resetMock()
-
-	// Common activities calls
-	s.env.OnActivity(
-		a.MpeCreationAcknowledgementActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Once()
-	s.env.OnActivity(
-		activities.FetchTracksInformationActivity,
-		mock.Anything,
-		initialTracksIDs,
-	).Return(initialTracksMetadata, nil).Once()
-
-	// Specific activities calls
-	s.env.OnActivity(
-		a.AcknowledgeDeletingTracksActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Never()
-
-	initialTracksFetched := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Equal(initialTracksMetadata, mpeState.Tracks)
-	}, initialTracksFetched)
-
-	deleteTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		unknownUserID := faker.UUIDHyphenated()
-		unknownDeviceID := faker.UUIDHyphenated()
-
-		s.emitDeleteTracksSignal(shared_mpe.NewDeleteTracksSignalArgs{
-			TracksIDs: initialTracksIDs,
-			UserID:    unknownUserID,
-			DeviceID:  unknownDeviceID,
-		})
-	}, deleteTracks)
-
-	checkDeletedTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Equal(initialTracksMetadata, mpeState.Tracks)
-	}, checkDeletedTracks)
-
-	s.env.ExecuteWorkflow(MpeRoomWorkflow, params)
-
-	s.True(s.env.IsWorkflowCompleted())
-	err := s.env.GetWorkflowError()
-	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
-}
-
-func (s *EditingPlaylistTestSuite) Test_AcknowledgeDeletionEvenWhenNoTracksWereDeleted() {
-	initialTracksIDs := []string{
-		faker.UUIDHyphenated(),
-		faker.UUIDHyphenated(),
-	}
-	params, roomCreatorDeviceID := s.getWorkflowInitParams(initialTracksIDs)
-	var a *activities_mpe.Activities
-
-	initialTracksMetadata := []shared.TrackMetadata{
-		{
-			ID:         initialTracksIDs[0],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-		{
-			ID:         initialTracksIDs[1],
-			Title:      faker.Word(),
-			ArtistName: faker.Name(),
-			Duration:   42000,
-		},
-	}
-
-	tick := 1 * time.Millisecond
-	resetMock, registerDelayedCallbackWrapper := s.initTestEnv()
-
-	defer resetMock()
-
-	// Common activities calls
-	s.env.OnActivity(
-		a.MpeCreationAcknowledgementActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Once()
-	s.env.OnActivity(
-		activities.FetchTracksInformationActivity,
-		mock.Anything,
-		initialTracksIDs,
-	).Return(initialTracksMetadata, nil).Once()
-
-	// Specific activities calls
-	s.env.OnActivity(
-		a.AcknowledgeDeletingTracksActivity,
-		mock.Anything,
-		mock.Anything,
-	).Return(nil).Once()
-
-	initialTracksFetched := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Equal(initialTracksMetadata, mpeState.Tracks)
-	}, initialTracksFetched)
-
-	deleteTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		tracksIDsToDelete := []string{
-			faker.UUIDHyphenated(),
-			faker.UUIDHyphenated(),
-		}
-
-		s.emitDeleteTracksSignal(shared_mpe.NewDeleteTracksSignalArgs{
-			TracksIDs: tracksIDsToDelete,
-			UserID:    params.RoomCreatorUserID,
-			DeviceID:  roomCreatorDeviceID,
-		})
-	}, deleteTracks)
-
-	checkDeletedTracks := tick * 200
-	registerDelayedCallbackWrapper(func() {
-		mpeState := s.getMpeState(shared_mpe.NoRelatedUserID)
-
-		s.Equal(initialTracksMetadata, mpeState.Tracks)
-	}, checkDeletedTracks)
-
-	s.env.ExecuteWorkflow(MpeRoomWorkflow, params)
-
-	s.True(s.env.IsWorkflowCompleted())
-	err := s.env.GetWorkflowError()
-	s.ErrorIs(err, workflow.ErrDeadlineExceeded, "The workflow ran on an infinite loop")
-}
-
-func TestEditingPlaylistTestSuite(t *testing.T) {
-	suite.Run(t, new(EditingPlaylistTestSuite))
+func TestAddTracksTestSuite(t *testing.T) {
+	suite.Run(t, new(AddTracksTestSuite))
 }
